Fix puppet4 service not being enabled on RHEL
Due to
https://github.com/theforeman/community-templates/commit/8eca3cf8665d6193b5b34a27fc118469a8598c3a#diff-8048fb7fe5c29050e20a1a592d25a6c0R35
none of the puppet agent service enabling code was being run after
puppet is installed from the PC1 repository.  After discussing on IRC,
it may be that at least for some platforms the service is enabled by
default on installation.  For both EL6 and EL7 this wasn't the case.

This commit fixes this and has been tested on Oracle Linux 6 and 7.

It should also do the correct thing on all recent versions of Fedora.
(systemd has been the default init system on Fedora since v15, and the
earliest version puppetlabs provides packages for is 20.)