Enable puppet runs in the installer
A custom fact $::is_installer can be used inside Puppet scripts to see if
the current run is inside the installer or not. This can be sometimes
required, e.g. for interacting with services that are not yet running
(e.g. cups).

With this change puppet actually applies configuration changes, and does
not only register, but also runs puppet directly inside the installer
if the host parameter `run-puppet-in-installer` is set.

The change is only tested for Ubuntu 14.04, 15.10 and 16.04, as it
requires a workaround for a puppet bug: the detection of systemd fails.

closes GH-273