Fixes #14845: Use the first partition in preseed (#271)
Instead of hardcoding /dev/sda, use the first disk that the debian
installer detects through the `list-devices disk` command. This can still
select the wrong device for the particular setup, but at least does not
install into SD card devices any more (we had that happening a couple of
times).