CoreOS: use integer comparison for version (#293)
* Use integer comparison

As CoreOS version numbers are now greater than 1000 the lexical comparison fails; convert to integer and compare using that.

* Split compound ExecStart into multiple lines

This was using bash `&&` to ensure consecutive success only commands. systemd 
has that built in so just use that directly.  This also tames the super long line to be much 
more readable.